Moved class to seperate file and fixed some ts warnings
